requirements.txt
django==4.2
django-admin-interface
django-crispy-forms
dj-database-url==0.5.0
pillow
djangorestframework
requests
django-paypal
paypal-checkout-serversdk
pymysql
django-jazzmin
django-extensions
pandas
openpyxl
psycopg2-binary
python-dotenv
